+++ src/parse/dreamberd-driver.hh
@@ -22,7 +22,7 @@ namespace parse {
 
     parse::location get_location() const;
 
-    ast::NumberExp* make_NumberExp(const parse::location& location, int value) const;
+    ast::NumberExp* make_NumberExp(const parse::location& location, float value) const;
     ast::BinaryOpExp* make_BinaryOpExp(const parse::location& location,
                                        ast::Exp* left,
                                        ast::BinaryOpExp::Oper operation,
